swap to connector classes extending base connector
