+++ packages/caliper-core/lib/worker/workload/workloadModuleInterface.js
@@ -24,7 +24,7 @@ class WorkloadModuleInterface {
      * @param {number} totalWorkers The total number of workers participating in the round.
      * @param {number} roundIndex The 0-based index of the currently executing round.
      * @param {object} roundArguments The user-provided arguments for the round from the benchmark configuration file.
-     * @param {BlockchainInterface} sutAdapter The adapter of the underlying SUT.
+     * @param {BlockchainConnector} sutAdapter The adapter of the underlying SUT.
      * @param {object} sutContext The custom context object provided by the SUT adapter.
      * @async
      */
