+++ packages/caliper-fabric/lib/connector-versions/v2/registrarHelper.js
